README: Update instructions to reflect the format change of search results.
Some time in the past week, Twitter changed the format of its search
results, making the `insert-json-v1.pl` script obsolete as the results
in question can now be handled with `insert-json.pl` like the rest.
This makes things rather simpler for us.

(Still keeping this script around in case they change their mind
around this, or if there are still other kinds of results returned in
the v1 format.)